Updated and separated  env
