+++ .env.example
@@ -1,0 +1,3 @@
+GOOGLE_CLIENT_ID=""
+GOOGLE_CLIENT_SECRET=""
+
