--- app/.env.example
@@ -1,5 +1,0 @@
-GOOGLE_CLIENT_ID=""
-GOOGLE_CLIENT_SECRET=""
-
-MONGODB_URL="mongodb+srv://username:password@cluster0.5iguurt.mongodb.net/?appName=Cluster0"
-
